requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.121.1
fastapi-cli==0.0.16
fastapi-cloud-cli==0.3.1
google-auth==2.40.3
google-genai==1.28.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
openpyxl==3.1.5
orjson==3.11.4
outcome==1.3.0.post0
packaging==25.0
pillow==10.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.33.2
Pygments==2.19.2
pyperclip==1.9.0
PySocks==1.7.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.4
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.6
rsa==4.9.1
selenium==4.34.2
sentry-sdk==2.43.0
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.49.3
tenacity==8.5.0
trio==0.30.0
trio-websocket==0.12.2
ttkbootstrap==1.14.2
typer==0.20.0
typing-inspection==0.4.1
typing_extensions==4.14.1
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
